+++ src/Ros3dToolbar.js
@@ -1,0 +1,7 @@
+/**
+* @author Sarah Osentoski - sarah.osentoski@us.bosch.com
+*/
+
+var ROS3DTOOLBAR= ROS3DTOOLBAR || {
+    REVISION : '1'
+};