+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 2013-01-01 - **r1**
- * Initial development of MYPROJECT [(fauthor)](https://github.com/fauthor/)
+ * Initial development of ros3dtoolbar [(Sarah Osentoski)](https://github.com/sosentos/)